make web (api) public: move internal to pkg
