+++ pkg/web/flavors.go
