+++ pkg/web/flavors_test.go
